Change your data starts to float

+++ 4youandme/src/main/java/org/fouryouandme/core/entity/yourdata/YourData.kt
@@ -4,5 +4,5 @@ data class YourData(
     val id: String,
     val title: String?,
     val body: String?,
-    val starts: Int?
+    val starts: Float?
 )
